Refactor: Add AppUtil class, UserRepository, and UserUseCase; update menu view and app title
