+++ lib/core/util/index.dart
@@ -1,3 +1,4 @@
+export 'package:bodymetrics/core/util/app_util.dart';
 export 'package:bodymetrics/core/util/constants/index.dart';
 export 'package:bodymetrics/core/util/custom/index.dart';
 export 'package:bodymetrics/core/util/dialogs/dialog_util.dart';
