+++ lib/core/widgets/buttons/text_icon_button.dart
@@ -1,4 +1,3 @@
-import 'package:bodymetrics/core/index.dart';
 import 'package:flutter/material.dart';
 
 class TextIconButton extends StatelessWidget {
