+++ lib/core/widgets/index.dart
@@ -1,6 +1,7 @@
 export 'package:bodymetrics/core/widgets/animated_image.dart';
 export 'package:bodymetrics/core/widgets/buttons/colorful_text_button.dart';
 export 'package:bodymetrics/core/widgets/buttons/custom_elevated.dart';
+export 'package:bodymetrics/core/widgets/buttons/text_icon_button.dart';
 export 'package:bodymetrics/core/widgets/cards/home_card.dart';
 export 'package:bodymetrics/core/widgets/charts/line_charts.dart';
 export 'package:bodymetrics/core/widgets/custom_app_bar.dart';
@@ -9,4 +10,3 @@ export 'package:bodymetrics/core/widgets/custom_error.dart';
 export 'package:bodymetrics/core/widgets/custom_text_field.dart';
 export 'package:bodymetrics/core/widgets/gradient_scaffold.dart';
 export 'package:bodymetrics/core/widgets/picked_widget.dart';
-export 'package:bodymetrics/core/widgets/buttons/text_icon_button.dart';
