+++ lib/feature/home/domain/repository/user_repository.dart
@@ -1,0 +1,14 @@
+part of '../../presentation/home.dart';
+
+@immutable
+final class UserRepository implements BaseUseCase<int?, AppModel> {
+  @override
+  Future<int?>? execute() {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<int?>? executeWithParams(AppModel params) {
+    throw UnimplementedError();
+  }
+}
