+++ lib/feature/home/domain/use_case/user_use_case.dart
@@ -1,0 +1,16 @@
+part of '../../presentation/home.dart';
+
+@immutable
+final class UserUseCase implements BaseUseCase<int?, AppModel> {
+  @override
+  Future<int?>? execute() {
+
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<int?>? executeWithParams(AppModel params) {
+
+    throw UnimplementedError();
+  }
+}
