+++ lib/feature/home/presentation/home_model.dart
@@ -9,6 +9,7 @@ mixin _HomeModel on TickerProviderStateMixin<Home>, _TitleMixin {
 
   late AnimationController _animatedListController;
   late AnimationController _animatedChartController;
+
   @override
   void initState() {
     _bottomTitle = _bottomTitlesWeek;
